Code cleaning Added presenters for mode cases Changed searching logic

+++ app/src/main/java/com/sargis/kh/apixu/models/favorite/CurrentWeatherDataModel.java
@@ -8,7 +8,7 @@ public class CurrentWeatherDataModel implements Serializable {
 
     public Long id;
 
-    public Long orderIndex;
+    public Integer orderIndex;
 
     public boolean isSelected;
